requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
numpy==1.18.2
pandas==1.0.3
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
smmap==3.0.1
urllib3==1.25.9
